fix: Refactored after merge from main
